+++ src/main/java/core/net/LigaAuswahlDialog.java
@@ -38,26 +38,22 @@ public class LigaAuswahlDialog extends JDialog implements ActionListener {
     private JRadioButton m_jrbLigaAndere = new JRadioButton(core.model.HOVerwaltung.instance().getLanguageString("AndereLiga"),
                                                             false);
     private int m_iLigaId = -2;
-
+    private int ownLeagueId;
     //~ Constructors -------------------------------------------------------------------------------
 
     /**
      * Creates a new LigaAuswahlDialog object.
      */
-    public LigaAuswahlDialog(JDialog owner, int seasonid) {
+    public LigaAuswahlDialog(JDialog owner, int seasonid, int leagueId) {
         super(owner,
               core.model.HOVerwaltung.instance().getLanguageString("Liga"),
               true);
 
+        ownLeagueId = leagueId;
         setDefaultCloseOperation(JDialog.DISPOSE_ON_CLOSE);
         initComponents(seasonid);
     }
 
-    //~ Methods ------------------------------------------------------------------------------------
-
-    /**
-     * Gibt die Ligaid zurück oder -1 für aktulleLigaid oder -2 für abbruch
-     */
     public final int getLigaID() {
         return m_iLigaId;
     }
@@ -65,13 +61,10 @@ public class LigaAuswahlDialog extends JDialog implements ActionListener {
     @Override
 	public final void actionPerformed(ActionEvent e) {
         if (e.getSource().equals(m_jbOk)) {
-            //Ist es eine Zahl
             if (m_jrbLigaAktuell.isSelected()) {
-                //Aktuelle Liga benutzen
-                m_iLigaId = HOVerwaltung.instance().getModel().getXtraDaten().getLeagueLevelUnitID();
+                m_iLigaId = ownLeagueId;
                 setVisible(false);
             }
-            //Andere Liga
             else if (m_jcbLiga.getSelectedItem() != null) {
                 m_iLigaId = parseInt(this, m_jcbLiga.getSelectedItem().toString(), false);
 
@@ -83,11 +76,7 @@ public class LigaAuswahlDialog extends JDialog implements ActionListener {
             setVisible(false);
         } else if ((e.getSource().equals(m_jrbLigaAktuell))
                    || (e.getSource().equals(m_jrbLigaAndere))) {
-            if (m_jrbLigaAndere.isSelected()) {
-                m_jcbLiga.setEnabled(true);
-            } else {
-                m_jcbLiga.setEnabled(false);
-            }
+            m_jcbLiga.setEnabled(m_jrbLigaAndere.isSelected());
         }
     }
 
@@ -96,7 +85,7 @@ public class LigaAuswahlDialog extends JDialog implements ActionListener {
         return core.db.DBManager.instance().getAllLigaIDs();
     }
 
-    private void initComponents(int seasonid) {
+    private void initComponents( int seasonid) {
         setContentPane(new ImagePanel(new GridLayout(4, 2, 4, 4)));
 
         JLabel label = new JLabel(core.model.HOVerwaltung.instance().getLanguageString("Season"));
@@ -115,7 +104,7 @@ public class LigaAuswahlDialog extends JDialog implements ActionListener {
         getContentPane().add(m_jrbLigaAktuell);
 
         //Platzhalter
-        label = new JLabel();
+        label = new JLabel(""+ ownLeagueId);
         getContentPane().add(label);
 
         m_jrbLigaAndere.setToolTipText(HOVerwaltung.instance().getLanguageString("tt_LigaDownload_Andere"));
@@ -127,7 +116,7 @@ public class LigaAuswahlDialog extends JDialog implements ActionListener {
         m_jcbLiga = new JComboBox(fillCB());
         m_jcbLiga.setToolTipText(HOVerwaltung.instance().getLanguageString("tt_LigaDownload_LigaID"));
         m_jcbLiga.setEnabled(false);
-        m_jcbLiga.setSelectedItem(Integer.valueOf(DBManager.instance().getLigaID4SaisonID(seasonid)));
+        m_jcbLiga.setSelectedItem(DBManager.instance().getLigaID4SaisonID(seasonid));
         m_jcbLiga.setEditable(true);
         getContentPane().add(m_jcbLiga);