+++ src/main/resources/release_notes.md
@@ -42,6 +42,7 @@
 ### League
 * fix download of power rating (not only one team per match) (#1293)
 * download all matches of latest league match day (#1373)
+* fix download of old league data (#1413)
   
 ### Youth
 * fix bug in potential calculation due to wrong isTop3 skill tagging (#1278)